Adk_Agent/requirements.txt
google-adk>=0.1.0
pandas>=1.5.0
requests>=2.28.0
pydantic>=1.10.0
